Solution to S1C02
A function to encode ASCII to hex was created, as well as a function to
compute the fixed XOR of the bytes in two given strings. Using the
previously created function to decode hex to ascii, the problem was
pretty straightforward. Decode the two input hex strings, compute their
fixed XOR and hex encode the result.